Protect the evaluation of the condition and the body of rules.
I'm doing this because, frankly, I don't trust my own
code (specifically, the partical and structural rules) to not contain
bugs that would lead to outright errors.  I wanted to provide a way of
evaluating possibly buggy rules in such a way that if an error occurs,
I'm notified and the rest of the rules are evaluated.