Deal with formulas and terms that could be equal (EQ, in fact) to symbolic attacks.
The approach, in the guise of NON-SYMBOLIC-ATTACK-TERM and
NON-SYMBOLIC-ATTACK-FORMULA, is unsatisfactory.  What's needed is a
more robust notion of language that can discriminate between formulas
and terms.  The current approach actually dies when playing the
dialogue game on "(and p (implies p p))", because "p" is considered a
term.  Ugh.

Well, I'm thankful that by treating the argumentation forms more
rigorously (that is, as rules on a par with the dialogue rules), these
problems are coming through clearly.