Ignore SIGNATURE in COMPOSITE-FORMULA?
This is unsatisfactory.  I use the signature in determining whether a
an arbitrary lisp object is an atomic formula or a formula, but *not*
when checking whether an arbitrary lisp object is a composite formula.

I could define COMPOSITE-FORMULA? as simply the negation of
ATOMIC-FORMULA?, but this doesn't feel right either, because that
seems to be a lot of computation.

Wait a second: does this function even do what I say it does?  It
doesn't check whether an *arbitrary* lisp object is a composite
formula -- it just checks whether a given lisp object is a non-null
list that begins with either NOT, OR, AND, etc.  So unsatisfactory.