First go at proving a means for playing classical dialogues.
I just delete rule D11 from the list of rules. (Rule E is also
excluded, as for vanilla intuitionistic games.)

(But is this really correct?  I just tried playing the game for
excluded middle, (OR P (NOT P)).  It goes like this:

0 P       (OR P (NOT P))
1 O [A,0] WHICH-DISJUNCT?
2 P [D,1] (NOT P)
3 O [A,2] P

For step 4, I'd like Proponent to respond by re-defending against the
attack.  But that violates rule D12 ("An attack can be answered at
most once").)