Simplify WITH-YES-OR-NO.
The name doesn't seem to be a good fit (macros beginning with WITH-
typically evaluate their bodies in new environments).  But it's clear
that we don't need to keep the value of the response around (it's
supposed to be just a symbol, Y, YES, N, or NO), and the forms to be
executed already depend on that response just by virtue of how the
macro is called.