Still trying to capture classical logic.
Deleting rule D11 alone doesn't do it: when playing the dialogue
for (OR P (NOT P)) we naturally violate D12: when the dialogue looks
like

  0 P       (OR P (NOT P))
  1 O [A,0] WHICH-DISJUNCT?
  2 P [D,1] (NOT P)
  3 O [A,2] P

and for move 4 we try to get Proponent to (re)defend against the
attack at move 1, we get:

  [D12] Attacks may be answered at most once.

So D12 is violated.  But we can't keep D11 and delete D12, either,
because the same move in the same dialogue violates D11: the most
recent open attack at step 4 is the attack at step 3 on (NOT
P).  (Attacks on negations are never closed; if D12 is present, one is
forced to respond immediately to attacks against one's assertion of a
negation.)

So what is classical logic here?

Do we even get intuitionstic logic?  Can Proponent win the game for,
say, the double negation of excluded middle?