Take the statements of rule D13 and rule E literally.
In the presence of Felscher's rule D00, play alterantes between
Proponent and Opponent, starting (move 0) with Proponent.  So "the
number of current move is odd" is equivalent to saying "it's
Opponent's turn".  And I used this equivalence when expressing the
rule condition for rules D13 and E.  Take the statements of the rule
literally, i.e., check that the current player is Opponent, not
whether the number of the current move is odd.  This might be needless
literalness, but perhaps there will be some dialogues where the
aforementioned equivalence no longer holds, but in which rules D13 and
E still apply.