Be more careful in WITH-ORIGINAL-STATEMENT.
We need to test whether there even is an original statement.  If not,
give null.  Without these tests, errors can be raised.  Here's one
example: consider the possibility that, say, Opponent chooses to
defend the first statement.  Since the move is a defense, we try to
hunt down the original statement, and die quickly because the initial
move isn't even an attack.  This is a crazy scenario, but it ought to
be handled nonetheless.