First stab at NEXT-MOVES.
It's pretty inefficient, I'd imagine.  The implementation is naive.
But this will probably do the job.  Hopefully, most of the silly
combinations of moves and attack/defense will quickly result in a rule
failure, and this function won't be as inefficient as I fear.

Calling NEXT-MOVES on fairly simple dialogues will be good stress
test.