Make sure that attacks and defenses alternate.
As I expected, having NEXT-MOVES around is a good stress test for the
various rules.  To my surprise, there was an obvious gap in the rules:
I never included a rules that attacks can never be against oneself,
and defenses are against the attacks of the other player, which are in
turn against the first players statements.  An embarrassing oversight!