Define GENERAL-SEARCH-FOR-BOTTOM.
The search idiom defined by GENERAL-SEARCH takes a node from a queue,
determines whether it is a goal node, and, when it is not a goal node,
expands it and enqueues the successors of the node.  When one is
searching simply for a node without successors, a simpler procedure is
available; GENERAL-SEARCH-FOR-BOTTOM implements it.  Without such a
function, carrying our a search for bottom is done in a somewhat
awkward way: either

1. for GOAL-TEST, carry out the calculation of successors until at
least one is found, thereby returning NIL; if no successors are found,
return T. The SUCCESSORS method then computes again a good deal of the
calculation that GOAL-TEST just computed; or

2. GOAL-TEST calculates the set of all successors of a node, and sets
the nodes SUCCESSORS slot to the newly computed value; the SUCCESSORS
method then simply tests whether a node is both expanded and has an
empty list of successors.

I don't know of any other alternatives. I dislike the second approach
because it feels like testing whether a node is a goal shouldn't
modify the state of the node.  The first approach is dispreffered
because of the possibly redundant calculation of the successors.