If one is careful, this seems to work.
At least in one round.  A LOT of functionality is still missing.

I've merged the dialogues and dialogue-site system definitions.  I was
running into too many subtle bugs having to do with the face that the
two packages were different.  Things like

  (eq <something> 'p)

would fail, because <something> would be in the DIALOGUE-SITE package;
even if its symbol name is equal to the symbol name of 'p (understood
as belonging to the DIALOGUES package), they would be different
symbols, so the EQ would be false.  D'oh.