Eliminate the signature editor (for now).
It seems like overkill to allow the possibility of editing the
signature in the middle of the game.  I used to think that that was
disallowed.  Why did I ever think that it *should* be allowed?  It's
not entirely useless to edit the signature mid-game, I suppose, but it
hardly seems like something to focus on now.