Add a bunch of new formulas.
Most of these are intuitionistically valid, but some are strictly classical.

The list (and most of the names) is culled from

  http://editthis.info/logic/The_Laws_of_Classical_Logic

This website is very sketchy, but the list is fine (though a handfule
of the names are unusual to me).  Thanks to Sara Uckelman for pointing
out this list to me.